kernel: Increase max swappiness limit to 250
Swappiness is a property for the Linux kernel that changes the balance between swapping out runtime
memory, as opposed to dropping pages from the system page cache. A low value means the kernel will try
to avoid swapping as much as possible where a higher value instead will make the kernel aggressively
try to use swap space.

For in-memory swap, like zram or zswap, as well as hybrid setups that have swap on faster devices
than the filesystem, values beyond 200 might be considered.

Some benchmark test on Galaxy A10:
-Random Access Read speed: 52.6 MB/s
-Random Access Write speed: 7.5 MB/s

Change-Id: I651171d81784a0a7376ce1618eb984cd00bfb285
Signed-off-by: Chatur27 <jasonbright2709@gmail.com>